[refactor] Re-arrange the `torii` crate.
