+++ test_js/config.json
@@ -1,0 +1,42 @@
+{
+    "TORII": {
+        "P2P_ADDR": "127.0.0.1:1337",
+        "API_URL": "127.0.0.1:8080",
+        "TELEMETRY_URL": "127.0.0.1:8081"
+    },
+    "SUMERAGI": {
+        "TRUSTED_PEERS": [
+            {
+                "address": "127.0.0.1:1337",
+                "public_key": "ed01207233bfc89dcbd68c19fde6ce6158225298ec1131b6a130d1aeb454c1ab5183c0"
+            }
+        ],
+        "BLOCK_TIME_MS": 100,
+        "COMMIT_TIME_MS": 100,
+        "TX_RECEIPT_TIME_MS": 100,
+        "MAX_FAULTY_PEERS": 0
+    },
+    "KURA": {
+        "INIT_MODE": "strict",
+        "BLOCK_STORE_PATH": "./blocks"
+    },
+    "LOGGER": {
+        "MAX_LOG_LEVEL": "DEBUG"
+    },
+    "BLOCK_SYNC": {
+        "GOSSIP_PERIOD_MS": 10000,
+        "BATCH_SIZE": 4
+    },
+    "PUBLIC_KEY": "ed01207233bfc89dcbd68c19fde6ce6158225298ec1131b6a130d1aeb454c1ab5183c0",
+    "PRIVATE_KEY": {
+        "digest_function": "ed25519",
+        "payload": "9ac47abf59b356e0bd7dcbbbb4dec080e302156a48ca907e47cb6aea1d32719e7233bfc89dcbd68c19fde6ce6158225298ec1131b6a130d1aeb454c1ab5183c0"
+    },
+    "GENESIS": {
+        "ACCOUNT_PUBLIC_KEY": "ed01207233bfc89dcbd68c19fde6ce6158225298ec1131b6a130d1aeb454c1ab5183c0",
+        "ACCOUNT_PRIVATE_KEY": {
+            "digest_function": "ed25519",
+            "payload": "9ac47abf59b356e0bd7dcbbbb4dec080e302156a48ca907e47cb6aea1d32719e7233bfc89dcbd68c19fde6ce6158225298ec1131b6a130d1aeb454c1ab5183c0"
+        }
+    }
+}
