+++ test_js/genesis.json
@@ -1,0 +1,98 @@
+{
+    "transactions": [
+        {
+            "isi": [
+                {
+                    "Register": {
+                        "object": {
+                            "Raw": {
+                                "Identifiable": {
+                                    "Domain": {
+                                        "id": {
+                                            "name": "wonderland"
+                                        },
+                                        "accounts": {},
+                                        "asset_definitions": {},
+                                        "metadata": {}
+                                    }
+                                }
+                            }
+                        }
+                    }
+                },
+                {
+                    "Register": {
+                        "object": {
+                            "Raw": {
+                                "Identifiable": {
+                                    "NewAccount": {
+                                        "id": {
+                                            "name": "alice",
+                                            "domain_id": {
+                                                "name": "wonderland"
+                                            }
+                                        },
+                                        "signatories": [
+                                            "ed01207233bfc89dcbd68c19fde6ce6158225298ec1131b6a130d1aeb454c1ab5183c0"
+                                        ],
+                                        "metadata": {}
+                                    }
+                                }
+                            }
+                        }
+                    }
+                },
+                {
+                    "Register": {
+                        "object": {
+                            "Raw": {
+                                "Identifiable": {
+                                    "AssetDefinition": {
+                                        "id": {
+                                            "name": "rose",
+                                            "domain_id": {
+                                                "name": "wonderland"
+                                            }
+                                        },
+                                        "value_type": "Quantity",
+                                        "metadata": {},
+                                        "mintable": true
+                                    }
+                                }
+                            }
+                        }
+                    }
+                },
+                {
+                    "Mint": {
+                        "object": {
+                            "Raw": {
+                                "U32": 13
+                            }
+                        },
+                        "destination_id": {
+                            "Raw": {
+                                "Id": {
+                                    "AssetId": {
+                                        "definition_id": {
+                                            "name": "rose",
+                                            "domain_id": {
+                                                "name": "wonderland"
+                                            }
+                                        },
+                                        "account_id": {
+                                            "name": "alice",
+                                            "domain_id": {
+                                                "name": "wonderland"
+                                            }
+                                        }
+                                    }
+                                }
+                            }
+                        }
+                    }
+                }
+            ]
+        }
+    ]
+}
